feature: asynchronous preview of the ogmarkup document
This patch focuses on making the preview rendering of an ogmarkup as
asynchronous as possible. A new version of pursuit-ogmarkup is leveraged
in order to do so. The main benefit of this approach is to have a
preview function which does not freeze the single thread of a web
application.